+++ CHANGELOG.md
@@ -56,6 +56,7 @@
 * nostr: require `fmt::Debug`, `Send` and `Sync` for `NostrSigner` ([Yuki Kishimoto])
 * nostr: enable support for `Tags::indexes` in `no_std` ([Yuki Kishimoto])
 * nostr: improve `RelayMessage` docs ([Yuki Kishimoto])
+* nostr: dedup tags in `EventBuilder::text_note_reply` and `EventBuilder::comment` ([Yuki Kishimoto])
 * database: add manual trait implementations for `BTreeCappedSet` ([Yuki Kishimoto]) 
 * database: replace LRU with custom memory cache for IDs tracking ([Yuki Kishimoto])
 * lmdb: use `async-utility` to spawn blocking tasks ([Yuki Kishimoto])